Create GenomeRow to reuse elements.
