+++ src/components/Actions/GenomeRow.module.scss
@@ -1,0 +1,8 @@
+.AllGenomesList {
+  gap: 0.25rem;
+  align-items: flex-start;
+  margin: 0;
+  padding: 0 2.5rem;
+  font-size: 0.75rem;
+  text-align: left;
+}
